Codereview comments
