+++ OneSignalSDK/onesignal/src/main/AndroidManifest.xml
@@ -92,7 +92,7 @@
         <activity android:name="com.onesignal.PermissionsActivity"
                   android:theme="@android:style/Theme.Translucent.NoTitleBar" />
 
-        <receiver android:name="com.onesignal.NotificationOpenedReceiver" />
+        <receiver android:name="com.onesignal.NotificationDismissReceiver" />
 
         <receiver android:name="com.onesignal.BootUpReceiver">
             <intent-filter>
@@ -107,7 +107,7 @@
         </receiver>
 
         <activity
-          android:name="com.onesignal.NotificationOpenedActivity"
+          android:name="com.onesignal.NotificationOpenedReceiver"
           android:noHistory="true"
           android:theme="@android:style/Theme.Translucent.NoTitleBar" />
