+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/NotificationDismissReceiver.java
@@ -1,6 +1,4 @@
-/**
- * Modified MIT License
- *
+/*
  * Copyright 2017 OneSignal
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -27,24 +25,14 @@
 
 package com.onesignal;
 
-import android.app.Activity;
+import android.content.BroadcastReceiver;
+import android.content.Context;
 import android.content.Intent;
-import android.os.Bundle;
 
-public class NotificationOpenedActivity extends Activity {
+public class NotificationDismissReceiver extends BroadcastReceiver {
 
-   @Override
-   protected void onCreate(Bundle savedInstanceState) {
-      super.onCreate(savedInstanceState);
-      NotificationOpenedProcessor.processFromContext(this, getIntent());
-      finish();
-   }
-
-   @Override
-   protected void onNewIntent(Intent intent) {
-      super.onNewIntent(intent);
-      NotificationOpenedProcessor.processFromContext(this, getIntent());
-      finish();
-   }
-
-}
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        NotificationOpenedProcessor.processFromContext(context, intent);
+    }
+}
